# src/routes/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Record_string.unknown_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
...
			"SdJwtVcRecord": {
				"$ref": "#/components/schemas/Record_string.unknown_"
			},
			"MdocRecord": {
				"$ref": "#/components/schemas/Record_string.unknown_"
			},
			"AuthorizeRequestCredentialOffer": {
				"properties": {
					"credentialOfferUri": {
						"type": "string"
					},
					"credentialsToRequest": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"credentialOfferUri",
					"credentialsToRequest"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RequestCredentialBody": {
				"properties": {
					"credentialOfferUri": {
						"type": "string"
					},
					"credentialsToRequest": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"authorizationCode": {
						"type": "string"
					},
					"codeVerifier": {
						"type": "string"
					},
					"txCode": {
						"type": "string"
					}
				},
				"required": [
					"credentialOfferUri",
					"credentialsToRequest"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResolveProofRequest": {
				"properties": {
					"proofRequestUri": {
						"type": "string"
					}
				},
				"required": [
					"proofRequestUri"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OpenId4VcIssuerRecord": {
				"$ref": "#/components/schemas/Record_string.unknown_",
				"description": "For OID4VC you need to expose metadata files. Each issuer needs to host this metadata. This is not the case for DIDComm where we can just have one /didcomm endpoint.\nSo we create a record per openid issuer/verifier that you want, and each tenant can create multiple issuers/verifiers which have different endpoints\nand metadata files"
			},
			"Logo": {
				"properties": {
					"uri": {
						"type": "string"
					},
					"alt_text": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": {}
			},
			"CredentialDisplay": {
				"properties": {
					"name": {
						"type": "string"
					},
					"locale": {
						"type": "string"
					},
					"logo": {
						"$ref": "#/components/schemas/Logo"
					}
				},
				"type": "object",
				"additionalProperties": {}
			},
			"ProofTypeConfig": {
				"properties": {
					"proof_signing_alg_values_supported": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"proof_signing_alg_values_supported"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.ProofTypeConfig_": {
				"properties": {},
				"additionalProperties": {
					"$ref": "#/components/schemas/ProofTypeConfig"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"CredentialDefinition": {
				"properties": {
					"type": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"type"
				],
				"type": "object",
				"additionalProperties": {}
			},
			"CredentialConfigurationDisplay": {
				"properties": {
					"name": {
						"type": "string"
					},
					"locale": {
						"type": "string"
					},
					"logo": {
						"$ref": "#/components/schemas/Logo"
					},
					"description": {
						"type": "string"
					},
					"background_color": {
						"type": "string"
					},
					"background_image": {
						"$ref": "#/components/schemas/Logo"
					},
					"text_color": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CredentialConfigurationSupportedWithFormats": {
				"properties": {
					"format": {
						"type": "string",
						"enum": [
							"jwt_vc_json"
						],
						"nullable": false
					},
					"scope": {
						"type": "string"
					},
					"cryptographic_binding_methods_supported": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"credential_signing_alg_values_supported": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"proof_types_supported": {
						"$ref": "#/components/schemas/Record_string.ProofTypeConfig_"
					},
					"credential_definition": {
						"$ref": "#/components/schemas/CredentialDefinition"
					},
					"display": {
						"items": {
							"$ref": "#/components/schemas/CredentialConfigurationDisplay"
						},
						"type": "array"
					}
				},
				"required": [
					"format",
					"credential_definition"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.CredentialConfigurationSupportedWithFormats_": {
				"properties": {},
				"additionalProperties": {
					"$ref": "#/components/schemas/CredentialConfigurationSupportedWithFormats"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"BatchCredentialIssuanceOptions": {
				"properties": {
					"batchSize": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"batchSize"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateIssuerRecordOptions": {
				"properties": {
					"display": {
						"items": {
							"$ref": "#/components/schemas/CredentialDisplay"
						},
						"type": "array"
					},
					"dpopSigningAlgValuesSupported": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"credentialConfigurationsSupported": {
						"$ref": "#/components/schemas/Record_string.CredentialConfigurationSupportedWithFormats_"
					},
					"batchCredentialIssuance": {
						"$ref": "#/components/schemas/BatchCredentialIssuanceOptions"
					}
				},
				"required": [
					"credentialConfigurationsSupported"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"OpenId4VcVerifierRecord": {
				"$ref": "#/components/schemas/Record_string.unknown_",
				"description": "For OID4VC you need to expos metadata files. Each issuer needs to host this metadata. This is not the case for DIDComm where we can just have one /didcomm endpoint.\nSo we create a record per openid issuer/verifier that you want, and each tenant can create multiple issuers/verifiers which have different endpoints\nand metadata files"
			}
		},
		"securitySchemes": {
			"apiKey": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			},
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "credo-controller",
		"version": "2.0.0",
		"description": "Rest endpoint wrapper for using your agent over HTTP",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {}
	},
	"paths": {
...
		"/openid4vc/holder/sd-jwt-vcs": {
			"get": {
				"operationId": "GetSdJwtCredentials",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SdJwtVcRecord"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get SdJwt type of credentials",
				"tags": [
					"oid4vc holders"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": []
			}
		},
		"/openid4vc/holder/mdoc-vcs": {
			"get": {
				"operationId": "GetMdocCredentials",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MdocRecord"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Fetch all mso mdoc credentials in wallet",
				"tags": [
					"oid4vc holders"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": []
			}
		},
		"/openid4vc/holder/authorization-request": {
			"post": {
				"operationId": "RequestAuthorizationForCredential",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"codeVerifier": {
											"type": "string"
										},
										"authorizationRequestUrl": {
											"type": "string"
										},
										"actionToTake": {
											"type": "string"
										}
									},
									"required": [
										"codeVerifier",
										"authorizationRequestUrl",
										"actionToTake"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Resolve a credential offer",
				"tags": [
					"oid4vc holders"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthorizeRequestCredentialOffer"
							}
						}
					}
				}
			}
		},
		"/openid4vc/holder/request-credential": {
			"post": {
				"operationId": "RequestCredential",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Initiates a token request, then requests credentials from issuer",
				"tags": [
					"oid4vc holders"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RequestCredentialBody"
							}
						}
					}
				}
			}
		},
		"/openid4vc/holder/resolve-proof-request": {
			"post": {
				"operationId": "ResolveProofRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Resolve a proof request",
				"tags": [
					"oid4vc holders"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResolveProofRequest"
							}
						}
					}
				}
			}
		},
		"/openid4vc/holder/accept-proof-request": {
			"post": {
				"operationId": "AcceptProofRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Accept a proof request",
				"tags": [
					"oid4vc holders"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResolveProofRequest"
							}
						}
					}
				}
			}
		},
		"/openid4vc/issuer": {
			"post": {
				"operationId": "CreateIssuer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OpenId4VcIssuerRecord"
								}
							}
						}
					}
				},
				"description": "Creates an issuer with issuer metadata.",
				"tags": [
					"oid4vc issuers"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			},
			"get": {
				"operationId": "GetIssuersByQuery",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/OpenId4VcIssuerRecord"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Query issuers by optional publicIssuerId.",
				"tags": [
					"oid4vc issuers"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "publicIssuerId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/openid4vc/issuer/{publicIssuerId}": {
			"put": {
				"operationId": "UpdateIssuerMetadata",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OpenId4VcIssuerRecord"
								}
							}
						}
					}
				},
				"description": "Updates issuer metadata for a given publicIssuerId.",
				"tags": [
					"oid4vc issuers"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "publicIssuerId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateIssuerRecordOptions"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetIssuer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OpenId4VcIssuerRecord"
								}
							}
						}
					}
				},
				"description": "Returns a specific issuer by publicIssuerId.",
				"tags": [
					"oid4vc issuers"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "publicIssuerId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/openid4vc/issuer/{issuerId}/metadata": {
			"get": {
				"operationId": "GetIssuerAgentMetaData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"description": "Returns metadata for a specific issuer.",
				"tags": [
					"oid4vc issuers"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "issuerId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/openid4vc/issuer/{id}": {
			"delete": {
				"operationId": "DeleteIssuer",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Deletes a specific issuer by record id.",
				"tags": [
					"oid4vc issuers"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/openid4vc/verifier": {
			"post": {
				"operationId": "CreateVerifier",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OpenId4VcVerifierRecord"
								}
							}
						}
					}
				},
				"description": "Create a new verifier and store the verifier record",
				"tags": [
					"oid4vc verifiers"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			},
			"get": {
				"operationId": "GetVerifiersByQuery",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/OpenId4VcVerifierRecord"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get verifiers by query",
				"tags": [
					"oid4vc verifiers"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "publicVerifierId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/openid4vc/verifier/{publicVerifierId}": {
			"put": {
				"operationId": "UpdateVerifierMetadata",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OpenId4VcVerifierRecord"
								}
							}
						}
					}
				},
				"description": "Update verifier metadata",
				"tags": [
					"oid4vc verifiers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "publicVerifierId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			},
			"get": {
				"operationId": "GetVerifier",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OpenId4VcVerifierRecord"
								}
							}
						}
					}
				},
				"description": "Get single verifier by ID",
				"tags": [
					"oid4vc verifiers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "publicVerifierId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/openid4vc/verifier/{verifierId}": {
			"delete": {
				"operationId": "DeleteVerifier",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Delete verifier by ID",
				"tags": [
					"oid4vc verifiers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "verifierId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}